refact: antd dashboard menu
