--- src/app/protected/components/ButtonAntd.tsx
@@ -1,8 +1,0 @@
-'use client';
-
-import { Button } from 'antd';
-import React from 'react';
-
-export default function ButtonAntd() {
-  return <Button type="primary">Meu Botão</Button>;
-}
